general.dart file removed
