--- frontend/lendme/lib/exceptions/exception.dart
@@ -1,12 +1,0 @@
-class DomainException implements Exception {
-
-  final String message;
-  final Object? cause;
-
-  DomainException(this.message, {this.cause});
-
-  @override
-  String toString() {
-    return '${runtimeType.toString()}($message)';
-  }
-}
