+++ frontend/lendme/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:lendme/exceptions/general.dart';
 import 'package:lendme/repositories/user_repository.dart';
 import 'package:lendme/screens/proxy.dart';
 import 'package:lendme/services/auth.dart';
@@ -8,6 +7,7 @@ import 'package:lendme/services/emulators.dart';
 import 'package:provider/provider.dart';
 import 'package:stream_transform/stream_transform.dart';
 
+import 'exceptions/exceptions.dart';
 import 'models/resource.dart';
 import 'models/user.dart';
