+++ frontend/lendme/lib/screens/auth/email_sign_in.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:lendme/components/loadable_area.dart';
-import 'package:lendme/exceptions/exception.dart';
+import 'package:lendme/exceptions/exceptions.dart';
 import 'package:lendme/services/auth.dart';
 import 'package:lendme/utils/ui/error_snackbar.dart';
